+++ pkg/resource/grafana-agent-secret/reconciler.go
@@ -1,4 +1,4 @@
-package promtailclient
+package grafanaagentsecret
 
 import (
 	"context"
@@ -19,15 +19,15 @@ import (
 )
 
 // Reconciler implements a reconciler.Interface to handle
-// Promtail client: extra promtail config about where and how to send logs
+// grafana-agent secret: extra secret which stores logging write credentials
 type Reconciler struct {
 	client.Client
 }
 
-// ReconcileCreate ensures promtail secret is created with the right credentials
+// ReconcileCreate ensures grafana-agent secret is created with the right credentials
 func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
-	logger.Info("promtailclient create")
+	logger.Info("grafana-agent-secret create")
 
 	// Retrieve secret containing credentials
 	var loggingCredentialsSecret v1.Secret
@@ -44,20 +44,20 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 	}
 
 	// Get desired secret
-	desiredPromtailClientSecret, err := GeneratePromtailClientSecret(lc, &loggingCredentialsSecret, lokiURL)
+	desiredGrafanaAgentSecret, err := GenerateGrafanaAgentSecret(lc, &loggingCredentialsSecret, lokiURL)
 	if err != nil {
-		logger.Info("promtailclient - failed generating auth config!", "error", err)
+		logger.Info("grafana-agent-secret - failed generating auth config!", "error", err)
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
 	// Check if secret already exists.
-	logger.Info("promtailclient - getting", "namespace", desiredPromtailClientSecret.GetNamespace(), "name", desiredPromtailClientSecret.GetName())
-	var currentPromtailClientSecret v1.Secret
-	err = r.Client.Get(ctx, types.NamespacedName{Name: desiredPromtailClientSecret.GetName(), Namespace: desiredPromtailClientSecret.GetNamespace()}, &currentPromtailClientSecret)
+	logger.Info("grafana-agent-secret - getting", "namespace", desiredGrafanaAgentSecret.GetNamespace(), "name", desiredGrafanaAgentSecret.GetName())
+	var currentGrafanaAgentSecret v1.Secret
+	err = r.Client.Get(ctx, types.NamespacedName{Name: desiredGrafanaAgentSecret.GetName(), Namespace: desiredGrafanaAgentSecret.GetNamespace()}, &currentGrafanaAgentSecret)
 	if err != nil {
 		if apimachineryerrors.IsNotFound(err) {
-			logger.Info("promtailclient not found, creating")
-			err = r.Client.Create(ctx, &desiredPromtailClientSecret)
+			logger.Info("grafana-agent-secret not found, creating")
+			err = r.Client.Create(ctx, &desiredGrafanaAgentSecret)
 			if err != nil {
 				return ctrl.Result{}, errors.WithStack(err)
 			}
@@ -66,25 +66,25 @@ func (r *Reconciler) ReconcileCreate(ctx context.Context, lc loggedcluster.Inter
 		}
 	}
 
-	if !needUpdate(currentPromtailClientSecret, desiredPromtailClientSecret) {
-		logger.Info("promtailclient up to date")
+	if !needUpdate(currentGrafanaAgentSecret, desiredGrafanaAgentSecret) {
+		logger.Info("grafana-agent-secret up to date")
 		return ctrl.Result{}, nil
 	}
 
-	logger.Info("promtailclient - updating")
-	err = r.Client.Update(ctx, &desiredPromtailClientSecret)
+	logger.Info("grafana-agent-secret - updating")
+	err = r.Client.Update(ctx, &desiredGrafanaAgentSecret)
 	if err != nil {
 		return ctrl.Result{}, errors.WithStack(err)
 	}
 
-	logger.Info("promtailclient - done")
+	logger.Info("grafana-agent-secret - done")
 	return ctrl.Result{}, nil
 }
 
 // ReconcileDelete - Not much to do here when a cluster is deleted
 func (r *Reconciler) ReconcileDelete(ctx context.Context, lc loggedcluster.Interface) (ctrl.Result, error) {
 	logger := log.FromContext(ctx)
-	logger.Info("promtailclient delete")
+	logger.Info("grafana-agent-secret delete")
 
 	return ctrl.Result{}, nil
 }