+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Configure grafana-agent config to grab Kubernetes Events and send them to Loki.
+- Create grafana-agent extra secret to store logging write credentials.
 - Prepare some stuff for CAPI.
 - Upgrade go dependencies.